Health check routes updated to be a single base route
