+++ src/app.js
@@ -2,7 +2,6 @@ import express from "express";
 import cors from "cors";
 import cookieParser from "cookie-parser";
 import healthCheckRouter from "./routes/healthCheck.routes.js";
-import databaseCheckRouter from "./routes/databaseCheck.routes.js";
 
 const app = express();
 
@@ -19,6 +18,5 @@ app.use(express.static("public"));
 app.use(cookieParser());
 
 app.use("/api/v1/health-check", healthCheckRouter);
-app.use("/api/v1/database-check", databaseCheckRouter);
 
 export { app };
