--- src/routes/databaseCheck.routes.js
@@ -1,9 +1,0 @@
-import express from "express";
-import { databaseCheck } from "../controllers/databaseCheck.controller.js";
-
-const router = express.Router();
-
-// database checking route
-router.route("/").get(databaseCheck);
-
-export default router;
