--- src/utils/ApiError.js
@@ -1,23 +1,0 @@
-class ApiError extends Error {
-  constructor(
-    statusCode,
-    message = "Something went wrong.",
-    errors = [],
-    stack = ""
-  ) {
-    super(message);
-    this.statusCode = statusCode;
-    this.data = null;
-    this.message = message;
-    this.success = false;
-    this.errors = errors;
-
-    if (stack) {
-      this.stack = stack;
-    } else {
-      Error.captureStackTrace(this, this.constructor);
-    }
-  }
-}
-
-export { ApiError };
